+++ regressor/src/utils/class_averages.txt
@@ -1,1 +1,1 @@
-{"car": {"count": 14385, "total": [21898.43999999967, 23568.289999999495, 55754.239999999765]}, "pedestrian": {"count": 2280, "total": [3998.8900000000003, 1576.6400000000049, 1974.090000000009]}, "cyclist": {"count": 893, "total": [1561.5099999999982, 552.850000000001, 1569.5100000000007]}, "truck": {"count": 606, "total": [1916.6199999999872, 1554.710000000011, 6567.400000000018]}}
+{"car": {"count": 14357, "total": [21964.250000000404, 23240.639999999898, 55878.37999999985]}, "pedestrian": {"count": 2207, "total": [3901.400000000007, 1385.6300000000035, 1805.2399999999939]}, "cyclist": {"count": 734, "total": [1264.9199999999987, 418.09999999999843, 1299.7800000000034]}, "truck": {"count": 488, "total": [1640.7499999999961, 1273.379999999998, 4491.929999999992]}}
